rut gon code

+++ project/core/src/com/mygdx/game/controller/HitBox.java
@@ -1,12 +1,12 @@
 package com.mygdx.game.controller;
 
 public class HitBox {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private int width;
     private int height;
 
-    public HitBox(int x, int y, int width, int height) {
+    public HitBox(float x, float y, int width, int height) {
         this.x = x;
         this.y = y;
         this.width = width;
